store/postgres: Migrate subgraph schemas one at a time
We detect which subgraphs need to be migrated to the latest subgraph
schema, and kick off that migration. Migration happens as part of
transact_block_operations, so that subgraph ingestion pauses while
migration is happening. The number of subgraphs that can migrate at any
given time is limited to 2 to make sure we do not starve the tokio worker
pool.

With this change, subgraphs must be explicitly created with
graph::store::postgres::create_subgraph_deployment. It is no longer
possible to create a subgraph by simply creating a
SubgraphDeploymentEntity.

Tests now use the split entities setup, rather than storing their data in
the public.entities table.