store/postgres: Have store users kick off subgraph migrations explicitly
Rather than silently tacking a subgraph migration onto
transact_block_operations, just report to callers if there is a pending
migration and have them call migrate_subgraph_deployment explicitly to kick
off the migration.

This also makes it possible to report migration process on the subgraph's
logger so that they are easier to surface to subgraph users.